Refactor and split health checks into dedicated services
